feat(vanadium): add updatesPerSecond to config and add other infrastructure

+++ vanadium/src/sandbox/kotlin/github/businessdirt/axite/vanadium/VanadiumSandbox.kt
@@ -7,6 +7,6 @@ class VanadiumSandbox : VanadiumAdapter {
     }
 
     override fun initialize() {}
-    override fun update(deltaTime: Float) {}
+    override fun update(deltaTime: Long) {}
     override fun shutdown() {}
 }
